+++ apps/citizen-func/src/functions/GetMessage/handler.ts
@@ -1,4 +1,5 @@
 import { Context } from "@azure/functions";
+import { BlobServiceWithFallBack } from "@pagopa/azure-storage-legacy-migration-kit";
 import { InternalMessageResponseWithContent } from "@pagopa/io-functions-commons/dist/generated/definitions/InternalMessageResponseWithContent";
 import { TagEnum as TagEnumPayment } from "@pagopa/io-functions-commons/dist/generated/definitions/MessageCategoryPayment";
 import { PaymentData } from "@pagopa/io-functions-commons/dist/generated/definitions/PaymentData";
@@ -38,7 +39,6 @@ import {
   Ulid,
 } from "@pagopa/ts-commons/lib/strings";
 import { withoutUndefinedValues } from "@pagopa/ts-commons/lib/types";
-import { BlobService } from "azure-storage";
 import * as express from "express";
 import * as A from "fp-ts/lib/Apply";
 import * as E from "fp-ts/lib/Either";
@@ -48,6 +48,7 @@ import * as B from "fp-ts/lib/boolean";
 import { flow, pipe } from "fp-ts/lib/function";
 
 import { ThirdPartyData } from "../../generated/definitions/ThirdPartyData";
+import { getContentFromBlob } from "../../utils/message-content";
 import {
   ThirdPartyDataWithCategoryFetcher,
   getOrCacheService,
@@ -190,7 +191,7 @@ const getErrorOrMaybeThirdPartyData = async (
 export function GetMessageHandler(
   messageModel: MessageModel,
   messageStatusModel: MessageStatusModel,
-  blobService: BlobService,
+  blobService: BlobServiceWithFallBack,
   serviceModel: ServiceModel,
   redisClientFactory: RedisClientFactory,
   serviceCacheTtl: NonNegativeInteger,
@@ -205,7 +206,7 @@ export function GetMessageHandler(
         fiscalCode,
         messageId as NonEmptyString,
       )(), // FIXME: decode instead of cast
-      messageModel.getContentFromBlob(blobService, messageId)(),
+      getContentFromBlob(blobService, messageId)(),
     ]);
 
     if (E.isLeft(errorOrMaybeDocument)) {
@@ -356,7 +357,7 @@ export function GetMessageHandler(
 export function GetMessage(
   messageModel: MessageModel,
   messageStatusModel: MessageStatusModel,
-  blobService: BlobService,
+  blobService: BlobServiceWithFallBack,
   serviceModel: ServiceModel,
   redisClientFactory: RedisClientFactory,
   serviceCacheTtl: NonNegativeInteger,